feat(MemMap): Adding support to MemMap
ARM-NONEABI-GCC does not allow allocation control with pragma
directives.
This is only possible with pragma keyword before a function/variable
definition which leads to too much code changes to to make MemMap
effective as library.

Either way, the MemMap library and python script to configure and
generate is left for reference only.

More details on:
https://www.autosar.org/fileadmin/user_upload/standards/classic/3-0/AUTOSAR_SWS_MemoryMapping.pdf