feat(TM4C123 and Windows): Adding support for GCC on windows
Adding support for TM4C123 and VSCode Cortex Debug with OpenOCD and ARM
GDB

Changes to linker script and startup code were needed.

Disabling Middlewares was also required, just keeping TivaWare Drivers
in place for making the current code compatible with TM4C123.

This is a quick and dirty solution, improvements are required.